+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 3.0.10 (2020-06-07)
+
+**Added**
+
+- feat: implement step setup/teardown hooks
+- feat: support alter response in teardown hooks
+
+**Fixed**
+
+- fix: ensure upload ready
+- fix: add ExtendJSONEncoder to safely dump json data with python object, such as MultipartEncoder
+
 ## 3.0.9 (2020-06-07)
 
 **Fixed**